% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_SFMIC.R
\name{DSC_SFMiC}
\alias{DSC_SFMiC}
\title{DSC_SFMIC}
\usage{
DSC_SFMiC(maxMiC = 30, m = 2, theta = 0.8, theta_adapt = 0.1,
  Theta = 0.8, thetaClass = 0.7, minNumSFMiC = 4, timeWindow = 10,
  P = 9, ts = 50, thresholdTemp = 100, k_temp = 25, minWeight = 10,
  silhouetteThreshold = 0, fr_threshold = 0.6, description = NULL)
}
\description{
DSC_SFMIC
}
